requirements.txt
biopython ver 1.81